+++ R/db-odbc-mssql.R
@@ -73,6 +73,8 @@
       substr        = sql_prefix("SUBSTRING"),
       ceil          = sql_prefix("CEILING"),
       ceiling       = sql_prefix("CEILING"),
+      # Casting as BIT converts the Integer result into an actual logical
+      # values TRUE and FALSE
       is.null = function(x){
         build_sql(
           "CASE WHEN ", x ," IS NULL THEN CAST(1 AS BIT) ELSE CAST(0 AS BIT) END"
@@ -81,6 +83,9 @@
         build_sql(
           "CASE WHEN ", x ," IS NULL THEN CAST(1 AS BIT) ELSE CAST(0 AS BIT) END"
       )},
+      # TRIM is not supported on MS SQL versions under 2017
+      # https://docs.microsoft.com/en-us/sql/t-sql/functions/trim-transact-sql
+      # Best solution was to nest a left and right trims.
       trimws = function(x){
         build_sql(
           "LTRIM(RTRIM(", x ,"))"